splitting plotting utils into utils and configs
